Begin sanitisation of inputs. (SQL)
Major modifications to SQLInterface (now using preparedStatements) and personDatabase (to work with the former)

Minor modifications to all other files.

+++ src/TOC19/CheckOut.java
@@ -61,7 +61,7 @@ public final class CheckOut
 //		++logicalSize;
 //		return 0;
 //	}
-	public final int addProduct(String name, long price, int barCode)
+	public final int addProduct(String name, long price, long barCode)
 	{
 		/** 
 		Class CheckOut: Method addProduct
